feat: move creation to factory so that tableBreakpoints can use it in the future
